MySite/work/requirements.txt
Django==2.0.3
django-bootstrap3==9.1.0
django-ckeditor==5.4.0
django-contrib-comments==1.8.0
django-js-asset==1.0.0
djangorestframework==3.7.7
Markdown==2.6.11
Pillow==5.0.0
Pygments==2.2.0
pytz==2018.3
